fix(PageResource) set tm_date as a request attribute when no session exist Refs: #30552 (#30777)
Pull Request Summary:

Historically, the tm_date (time machine date) has been passed as a
session attribute. Many components of the PageResource rely heavily on
session data. However, with the transition of timeMachineDate to a
request parameter, and considering that the endpoint can now be consumed
via curl using a Bearer token—where there is insufficient information to
construct or restore a session—we needed to ensure that this data could
be propagated to the internal layers of the API.

These changes introduce support for handling the timeMachineDate as a
request parameter, enabling its use even when no session exists.
Additionally, this update lays the groundwork for supporting a broader
range of dates for capturing the publishDate.