fix(uve): Experiments and UVE start a loop trying to resolve queryParams (#30800)
This pull request includes changes to the `dot-experiments-shell`
component and its corresponding tests to handle query parameters and
improve test coverage. The most important changes include adding the
`ActivatedRoute` dependency, updating the component to remove specific
query parameters, and enhancing the test suite to simulate and verify
these changes.

Enhancements to query parameter handling:

*
[`core-web/libs/portlets/dot-experiments/portlet/src/lib/dot-experiments-shell/dot-experiments-shell.component.ts`](diffhunk://#diff-ff8a6883a47be6d7ea576fb1ca835633f6ac5c922a83d5ec39a9af39cdcff264L2-R2):
Added `ActivatedRoute` dependency and updated the
`removeVariantQueryParams` method to remove `mode`, `variantName`, and
`experimentId` query parameters if they are present.
[[1]](diffhunk://#diff-ff8a6883a47be6d7ea576fb1ca835633f6ac5c922a83d5ec39a9af39cdcff264L2-R2)
[[2]](diffhunk://#diff-ff8a6883a47be6d7ea576fb1ca835633f6ac5c922a83d5ec39a9af39cdcff264L11-R31)
[[3]](diffhunk://#diff-ff8a6883a47be6d7ea576fb1ca835633f6ac5c922a83d5ec39a9af39cdcff264R43)

Improvements to test coverage:

*
[`core-web/libs/portlets/dot-experiments/portlet/src/lib/dot-experiments-shell/dot-experiments-shell.component.spec.ts`](diffhunk://#diff-76e44c7376ea0d0331e161dc38be1e049e1e464f28d040dbee385382192f6a87L18-R23):
Added `queryParams` to the `ActivatedRouteMock` and created the
`overrideRouteSnashot` function to simulate query parameter changes.
Enhanced the test suite to verify the removal of specific query
parameters and ensure no navigation occurs when they are not present.
[[1]](diffhunk://#diff-76e44c7376ea0d0331e161dc38be1e049e1e464f28d040dbee385382192f6a87L18-R23)
[[2]](diffhunk://#diff-76e44c7376ea0d0331e161dc38be1e049e1e464f28d040dbee385382192f6a87R33-R53)
[[3]](diffhunk://#diff-76e44c7376ea0d0331e161dc38be1e049e1e464f28d040dbee385382192f6a87R75-R115)


### Screenshot


https://github.com/user-attachments/assets/cdb80d90-e22a-4cdf-a8ed-4748475d40a2