feat(content analytics) fixes #30521 : Allow users to pass down simple Strings to query for Content Analytics data (#30869)
### Proposed Changes
* After discussion with the team, here's the new data format for the
simple String query. By default, the `request` scheme is prepended to
the appropriate terms so users won't have to repeat so many times:
```json
{
    "measures": "count,totalSessions",
    "dimensions": "host,whatAmI,url",
    "timeDimensions": "createdAt,day:Last month",
    "filters": "totalRequest gt 0,whatAmI contains PAGE||FILE",
    "order": "count asc,createdAt asc",
    "limit": 50,
    "offset": 0
}
```
If you want to set a specific scheme, just pass it down in the JSON body
like this:
```json
{
    "scheme": "YOUR-SCHEME-NAME-HERE",
     ...
     ...
}
```
The following must be taken into account when putting a query together:
  * Measures: Values are separated by commas: `count,totalSessions`
  * Dimensions: Values are separated by commas: `host,whatAmI,url`
* Time Dimensions: Values are separated by a comma: `createdAt,day:Last
month` . The second parameter 'day' -- the "granularity" parameter -- is
optional, and separated from the date range by a colon.
* Filters: Values are separated by commas: `totalRequest gt 0,whatAmI
contains PAGE||FILE` . In the `contains` clause, values are separated by
`||`.
  * Order: Values are separated by comma: `count asc,createdAt asc`
  * Limit: Value is provided as is: `50`
  * Offset: Value is provided as is: `0`

This PR fixes: #30521