fix(FTM): prevent editing in headless on preview mode (#30895)
This pull request introduces a new feature to handle preview mode across
various parts of the application. The changes involve adding checks for
preview mode and modifying the behavior based on whether the application
is running in preview mode.

### Preview Mode Handling:

*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/store/features/editor/withEditor.ts`](diffhunk://#diff-86e692578757ed7f4f6cba5d0aeb07641312f3b17885825d1a45987153ae87f0L117-R120):
Added a check for preview mode and modified the `isPageReady` and
`isLoading` variables to account for preview mode.
[[1]](diffhunk://#diff-86e692578757ed7f4f6cba5d0aeb07641312f3b17885825d1a45987153ae87f0L117-R120)
[[2]](diffhunk://#diff-86e692578757ed7f4f6cba5d0aeb07641312f3b17885825d1a45987153ae87f0L134-L135)
*
[`core-web/libs/sdk/client/src/lib/editor/sdk-editor.ts`](diffhunk://#diff-da7281c94a8714f497eefea7843cb58b1b25aec7513e621a0214694697ac4034R13):
Introduced the `isPreviewMode` function and updated the `isInsideEditor`
function to return `false` if in preview mode.
[[1]](diffhunk://#diff-da7281c94a8714f497eefea7843cb58b1b25aec7513e621a0214694697ac4034R13)
[[2]](diffhunk://#diff-da7281c94a8714f497eefea7843cb58b1b25aec7513e621a0214694697ac4034R113-R118)
*
[`core-web/libs/sdk/client/src/lib/utils/page/common-utils.ts`](diffhunk://#diff-b7ec297f0f1414686b0318121839e205f7d0c0ad6465cfae9d41a141d0831ff9R65-R76):
Added the `isPreviewMode` function to check if the application is
running in preview mode based on URL parameters.
*
[`dotCMS/src/main/webapp/html/js/editor-js/sdk-editor.js`](diffhunk://#diff-96d7c3f8039d9f4a1bd8a90496f8f06c3e89a4289b8d35c7d8a4941eacbfb07dL1-R1):
Updated the editor initialization script to include a check for preview
mode.

### Videos

#### Angular


https://github.com/user-attachments/assets/ec5dd9bd-24db-4185-ab31-421009e99c85

#### ReactJS


https://github.com/user-attachments/assets/23dc8254-080a-4213-9110-d4fb38f40287