Fix (Core): Changed job manager initialization to the InitServlet and processors discovery using CDI (#30898)
This pull request introduces several changes to improve the discovery
and registration of `JobProcessor` implementations using CDI (Contexts
and Dependency Injection).
The key changes include the addition of a new class for discovering job
processors, modifications to the job processor factory to utilize CDI,
and updates to the job queue manager to register discovered processors
automatically also, the initialization of the Job Manager it is now done
in the InitServlet.

### Improvements to Job Processor Discovery and Registration:

*
[`dotCMS/src/main/java/com/dotcms/jobs/business/api/JobProcessorDiscovery.java`](diffhunk://#diff-796e7503dcb4b03e85602c13708c084a19b3b5555bb5bbc46667ebabca1dc9dfR1-R88):
Introduced a new class `JobProcessorDiscovery` to discover all classes
implementing the `JobProcessor` interface using CDI. This class includes
methods to validate the scope of beans and handle errors during
discovery.

*
[`dotCMS/src/main/java/com/dotcms/jobs/business/api/JobProcessorFactory.java`](diffhunk://#diff-3da5d41bc50c5dc26ca6e83f9fd9364856ad73503723a820efb43231b75155e0R3-L31):
Modified the `JobProcessorFactory` to first attempt to get the processor
from the CDI context and fall back to direct instantiation if that
fails. This change ensures that job processors are created using CDI
when possible.

*
[`dotCMS/src/main/java/com/dotcms/jobs/business/api/JobQueueManagerAPIImpl.java`](diffhunk://#diff-c092f8af2f800c0ca84f2c2f4aed0af60d0ed488e20fa32992967e154064f561L130-R145):
Updated the `JobQueueManagerAPIImpl` class to include a
`JobProcessorDiscovery` instance and register discovered processors
automatically during initialization. Added methods to register job
processors with and without specifying a queue name.
[[1]](diffhunk://#diff-c092f8af2f800c0ca84f2c2f4aed0af60d0ed488e20fa32992967e154064f561L130-R145)
[[2]](diffhunk://#diff-c092f8af2f800c0ca84f2c2f4aed0af60d0ed488e20fa32992967e154064f561L150-R156)
[[3]](diffhunk://#diff-c092f8af2f800c0ca84f2c2f4aed0af60d0ed488e20fa32992967e154064f561R170-R172)
[[4]](diffhunk://#diff-c092f8af2f800c0ca84f2c2f4aed0af60d0ed488e20fa32992967e154064f561R254-R266)

### Removal of Deprecated Scanner:

*
[`dotCMS/src/main/java/com/dotcms/jobs/business/api/JobProcessorScanner.java`](diffhunk://#diff-63b494f914b5a884030a58cfa5de7c4410cc1fa7bf56bdf9ac6eb50e9cbc4a49L1-L66):
Removed the `JobProcessorScanner` class, which used Jandex to scan for
job processors. This functionality is now replaced by the new
`JobProcessorDiscovery` class.

### Updates to Job Processor Implementations:

* Added `@Dependent` scope to various job processor implementations to
ensure they are correctly managed by CDI:
  * `FailSuccessJob`
* `ImportContentletsProcessor`
[[1]](diffhunk://#diff-f5dc237cd813ef4bac45f0d238379d561192b07c9c1dd7e03619cde2ee3566adR48)
[[2]](diffhunk://#diff-f5dc237cd813ef4bac45f0d238379d561192b07c9c1dd7e03619cde2ee3566adR77)
  * `LargeFileReader`

These changes enhance the flexibility and reliability of job processor
discovery and instantiation in the system, leveraging CDI to manage
dependencies and lifecycle.