chore(edit-content): enhance dialog and table functionality (#30901)
### Parent Issue

https://github.com/dotCMS/core/issues/30523

### Proposed Changes

This pull request includes changes to the
`dot-select-existing-content.component.html` and
`dot-select-existing-content.component.ts` files to improve the user
interface and functionality of the content selection dialog. The most
important changes include adding a header template to the dialog, making
the dialog non-draggable, and introducing a method to check if an item
is selected.

Improvements to user interface:

*
[`dot-select-existing-content.component.html`](diffhunk://#diff-ba794904cc95635d9efca07b7e2bfeb948b5b0bb0c8e1e97c777c615398ba011R4-R22):
Added a header template to the dialog with a search icon and title.
*
[`dot-select-existing-content.component.html`](diffhunk://#diff-ba794904cc95635d9efca07b7e2bfeb948b5b0bb0c8e1e97c777c615398ba011R4-R22):
Made the dialog non-draggable and updated the style to include a height
property.

Functionality enhancements:

*
[`dot-select-existing-content.component.html`](diffhunk://#diff-ba794904cc95635d9efca07b7e2bfeb948b5b0bb0c8e1e97c777c615398ba011L79-R89):
Added a conditional class to highlight selected rows in the table.
*
[`dot-select-existing-content.component.ts`](diffhunk://#diff-14d3d666b0a074fe26436013ebf4abe96e518e28e9799b85b393d14b82ac528bR115-R123):
Introduced the `checkIfSelected` method to determine if an item is
selected.

### Checklist
- [x] Tests
- [x] Translations
- [x] Security Implications Contemplated (add notes if applicable)

### Screenshots

- The title should be "Search" and not "Select existing content".

<img width="1098" alt="Screenshot 2024-12-09 at 8 45 05 PM"
src="https://github.com/user-attachments/assets/1bacf286-8f9f-488d-b838-db292794663b">

- If the user make a search, close the dialog and open it again the
search input and the entries button are not clean, keeps the state.


https://github.com/user-attachments/assets/b92a8307-c43c-4f4c-b14f-fb3210c91b6a

- Following patterns, Dialog should keep the same height. when the user
perform a search with a couple of rows the height change


https://github.com/user-attachments/assets/286c9e15-29d3-4de4-ac7a-b9972928a973


- Form figma: The selected items will have an active state background.
- Dialog should not be draggable.


https://github.com/user-attachments/assets/767a049a-a713-4365-87b1-44c2c8f648ba