fix(uve) variant retrieval in contentSelector #30727 (#30933)
### Proposed Changes
- Added a new `variantName` property to `ContentSelector.js` to handle
variant-specific content searches.
- Updated `ng-contentlet-selector.jsp` to pass the `variantName`
parameter to `ContentSelector`.

### Checklist
- [x] Tests
- [x] Translations
- [x] Security Implications Contemplated (add notes if applicable)

### Additional Info
**Problem Statement:**  
When users create content for a specific variant in dotCMS, it does not
appear in the list when attempting to add a new piece of content to the
same variant. This issue was introduced in [PR
#30556](https://github.com/dotCMS/core/pull/30556), where
`window.sessionStorage` was no longer used to store the `variantName`.
The `ContentSelector.js` was still relying on this outdated
sessionStorage item, causing it to always default to `'DEFAULT'` for the
`variantName`.

### Video

https://github.com/user-attachments/assets/f527e92c-78f5-455b-9a3f-a02b7b4fe053