Issue 31023 lts designation (#31027)
This pull request includes updates to GitHub Actions workflows and
action configurations to enhance the automation processes for building,
testing, and releasing the project. The most important changes include
modifying the Docker image version handling, adding a new workflow for
LTS auto-run tests, and updating the release version validation.

### Workflow and Action Configuration Updates:

*
[`.github/actions/core-cicd/maven-job/action.yml`](diffhunk://#diff-55f17c9986a482bcfb132dc7b03fa4d1f89e17866d0dfa74cc6f41737c51f4adL244-R244):
Updated the Docker image version handling to use the `inputs.version`
parameter instead of a hardcoded version.

*
[`.github/workflows/cicd_5-lts.yml`](diffhunk://#diff-53c794c7312739238cb8a5e653444bbfb167a03cc4b3a90605130e208ed65f2cR1-R71):
Added a new GitHub Actions workflow configuration for running automated
tests on long-term support (LTS) branches. This workflow includes jobs
for initialization, building, testing, and finalizing with concurrency
control.

### Release Version Validation Updates:

*
[`.github/workflows/legacy-release_maven-release-process.yml`](diffhunk://#diff-9b645ac0f2192e1b22981f4186fbdc7fa0080f097f1ad0545f27f9250ed4410eL50-R50):
Updated the release version validation regex to allow for one or two
digits in the LTS version suffix. This change affects the validation of
inputs and the determination of whether a release is an LTS version.
[[1]](diffhunk://#diff-9b645ac0f2192e1b22981f4186fbdc7fa0080f097f1ad0545f27f9250ed4410eL50-R50)
[[2]](diffhunk://#diff-9b645ac0f2192e1b22981f4186fbdc7fa0080f097f1ad0545f27f9250ed4410eL82-R82)