fix(rest-api) content import postman flakey test #31005 (#31015)
### Proposed Changes
* Changed the scope of the variables `startTime` and `retryCount` from
`pm.environment` to `pm.collectionVariables` for better management and
scoping within Postman collections.
* Updated the test for canceling the content import job to use a larger
file to increase the chances of the job being in cancelable state when
the cancel request is made.

### Checklist
- [x] Tests
- [x] Translations
- [x] Security Implications Contemplated (add notes if applicable)

### Additional Info
These changes ensure that the retry logic and time tracking variables
are isolated to the collection level, avoiding potential conflicts with
environment-wide settings. Additionally, the larger file used for the
cancel job test makes it more likely to simulate a real-world scenario
where the job is still active when the cancel request is issued.