feat(deployment): Add shutdown-on-startup and skip-upgrade properties to support k8s init container use (#31003) (#31047)
### Proposed Changes
* Add a FinalStartupServlet servlet that initializes after all other
servlets that are set to run on startup. This will check for the
shutdown-on-startup config property (DOT_SHUTDOWN_ON_STARTUP as an env
var) default false, and if true the server will shut down with exit code
0 just before accepting connections. All startup initialization and
upgrade tasks should have been completed. This can be used to run as an
Init Container in K8s
* Add a skip-upgrade config ( DOT_SKIP_UPGRADE ) default false. If set
to true the DB startup tasks will not be attempted to be run. This can
be used in a k8s pod replica when it is known that the upgrade tasks
have been handled by the init container. There may not be much of a
downside of leaving this off as it should detect the work has already
been done, but this may help to reduce some unecessary work and also
distinguish if there are some tasks that need to be run per instance vs
per cluster/db.
* Add and fix some logging helping to distinguish the location of
startup steps between Servlets and ServletListeners.

### Testing
At this stage a basic test that the server does shut down with exit 0
after all startup tasks when DOT_SHUTDOWN_ON_STARTUP=true is passed, and
the fact that with the defaults, it does not impact any existing
behavior is required. This will be further tested when implementing the
init container that will make use of this flag.