chore(UVE): Hide Workflow Action Button on Variants (#31080)
This pull request introduces changes to the `core-web` library,
specifically within the `edit-ema` portlet, to add functionality for
displaying workflow actions based on certain conditions. The most
important changes include modifications to the `DotUveToolbarComponent`
and its related tests, as well as updates to the `withUVEToolbar` store
feature.

### Enhancements to `DotUveToolbarComponent`:

* Modified the HTML template to conditionally display the
`dot-uve-workflow-actions` component based on the `$showWorkflowActions`
method.
* Added the `$showWorkflowActions` observable to the component class to
determine whether to show workflow actions.
* Fix height on preview mode

### Updates to `DotUveToolbarComponent` tests:

* Added a new signal `$showWorkflowsActions` to the `baseUVEState` in
the component's spec file.
* Updated the test suite to include a test case for ensuring the
`dot-uve-workflow-actions` component is not present when
`$showWorkflowsActions` is set to false.

### Changes to `withUVEToolbar` store feature:

* Imported `UVE_MODE` and `DEFAULT_VARIANT_ID` for use in the store
feature.
* Added a new computed property `$showWorkflowsActions` to determine
whether to show workflow actions based on the editor mode and variant
ID.
* Added test cases for the `$showWorkflowsActions` property to ensure
correct behavior based on different conditions.

## Screenshots


https://github.com/user-attachments/assets/7f66d8be-2936-43d9-9f8d-495d3250d51d