feat(sdk): Added validation to blocks prop on BlockEditorRenderer (#31070)
https://github.com/user-attachments/assets/1753a058-b9af-4219-a13f-bbac1b716af1



This pull request introduces several changes to the
`BlockEditorRenderer` component and its related files to improve
validation and error handling for the `blocks` prop. The most important
changes include adding a new validation utility, updating the component
state to handle validation errors, and adding corresponding tests.

### Enhancements to `BlockEditorRenderer` component:

*
[`core-web/libs/sdk/react/src/lib/components/BlockEditorRenderer/BlockEditorRenderer.tsx`](diffhunk://#diff-c9fe945c43e268578d05f50d2302d2ec5caa376720ce5316472bf7663e52cefdL1-R1):
Added state management for validation errors and implemented a new
effect to validate the `blocks` prop using the `isValidBlocks` utility.
If the `blocks` prop is invalid, an error message is logged and
displayed conditionally based on the editor context.
[[1]](diffhunk://#diff-c9fe945c43e268578d05f50d2302d2ec5caa376720ce5316472bf7663e52cefdL1-R1)
[[2]](diffhunk://#diff-c9fe945c43e268578d05f50d2302d2ec5caa376720ce5316472bf7663e52cefdR10)
[[3]](diffhunk://#diff-c9fe945c43e268578d05f50d2302d2ec5caa376720ce5316472bf7663e52cefdR33-R37)
[[4]](diffhunk://#diff-c9fe945c43e268578d05f50d2302d2ec5caa376720ce5316472bf7663e52cefdR62-R81)
[[5]](diffhunk://#diff-c9fe945c43e268578d05f50d2302d2ec5caa376720ce5316472bf7663e52cefdR115-R143)

### New validation utility:

*
[`core-web/libs/sdk/react/src/lib/utils/utils.ts`](diffhunk://#diff-24fbd19d63c599f72373e849029ab266457a7c614ab67aca3530d78003506865R2):
Introduced the `isValidBlocks` utility function to validate the
structure of the `blocks` object, ensuring it has the correct type and
content properties.
[[1]](diffhunk://#diff-24fbd19d63c599f72373e849029ab266457a7c614ab67aca3530d78003506865R2)
[[2]](diffhunk://#diff-24fbd19d63c599f72373e849029ab266457a7c614ab67aca3530d78003506865R99-R113)

### Updates to `Block` interface:

*
[`core-web/libs/sdk/react/src/lib/models/blocks.interface.ts`](diffhunk://#diff-a445c20df92ec29029618009247f1300349020a2509376a1975ecfc886212414R11):
Added a `type` property to the `Block` interface to support the new
validation logic.

### Test enhancements:

*
[`core-web/libs/sdk/react/src/lib/components/BlockEditorRenderer/BlockEditorRenderer.spec.tsx`](diffhunk://#diff-00246bf45c5d583e9f0929525d681adc61a351ee44cad8d121a3ef0413be000bR13):
Added tests to verify that the `BlockEditorRenderer` component logs an
error and conditionally renders an error message when the `blocks` prop
is invalid.
[[1]](diffhunk://#diff-00246bf45c5d583e9f0929525d681adc61a351ee44cad8d121a3ef0413be000bR13)
[[2]](diffhunk://#diff-00246bf45c5d583e9f0929525d681adc61a351ee44cad8d121a3ef0413be000bR56-R75)

---------

Co-authored-by: Kevin Davila <kfariid@gmail.com>