fix(UVE): Add correct label to bookmark on old toolbar  (#31114)
This pull request includes changes to the `DotEmaBookmarksComponent` to
adjust the label behavior based on the preview mode and to add
corresponding test cases. The most important changes include updating
the label logic and adding tests to ensure the correct label is
displayed in different preview modes.

Label logic update:

*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/edit-ema-editor/components/dot-uve-toolbar/components/dot-ema-bookmarks/dot-ema-bookmarks.component.html`](diffhunk://#diff-73e6c9911b8aabc7460882885c63c407a12739d276a9cfbf8e68a159dbc0f6bbL9-R9):
Modified the label to be `editpage.toolbar.bookmark` when not in preview
mode.

Test cases addition:


*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/edit-ema-editor/components/dot-uve-toolbar/components/dot-ema-bookmarks/dot-ema-bookmarks.component.spec.ts`](diffhunk://#diff-0f88ad4304e13721edc116c3da29e02ab5e79004a21b2cbdab6e9c57e7ca772dR128-R137):
Added a test to check that the label is present when preview mode is
false.
*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/edit-ema-editor/components/dot-uve-toolbar/components/dot-ema-bookmarks/dot-ema-bookmarks.component.spec.ts`](diffhunk://#diff-0f88ad4304e13721edc116c3da29e02ab5e79004a21b2cbdab6e9c57e7ca772dR148-R157):
Added a test to check that the label is not present when preview mode is
true.

## Screenshots
<img width="409" alt="Screenshot 2025-01-13 at 5 30 34 PM"
src="https://github.com/user-attachments/assets/b77ae2e3-e0a1-4dc1-8ef6-b684df85d835"
/>