Issue 30126 wrong message when importing (#31055)
When importing content with a relationship field, the system did not
verify whether the value was a valid Lucene query or an identifier. This
led to keep processing of the field, ultimately resulting in an
incorrect error message.

A check has been added to ensure proper validation. 

Now, the import process will display an appropriate error message.

<img width="718" alt="Captura de pantalla 2025-01-03 a la(s) 5 31 34 p
 m"
src="https://github.com/user-attachments/assets/3f2f07a9-cc73-4e1e-9e25-6975b20710e0"
/>

Additionally, a new function was added to UtilMethods to validate
whether a string is a Lucene query.

---------

Co-authored-by: erickgonzalez <erick.gonzalez@dotcms.com>