fix(FTM) Homologous output Refs: #31061 (#31072)
### Proposed Changes
* This is quite simple. A cache invalidation was needed when passing FTM
params this was causing the output on the contents and the rendered bits
to diverge
* Additionally while updating the FTM karate test I discovered a
discrepancy between the rendered HTML and the contentlet items returned
within the container at the JSON level. A vtl code was used to render
the traditional TM content retrieved by the $dotcontent.pullPerPage view
tool was affecting the rendering logic. The VTL Logic was removed to
allow showing working versions of the content