fix(UVE): Unauthorized requests error due to Cookie Rejection on Firefox  (#31148)
This pull request includes changes to enhance the security and
functionality of the `edit-ema-editor` component. The most important
changes include adding a sandbox attribute to the iframe and modifying
the iframe URL generation to avoid caching issues.

Enhancements to security and functionality:

*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/edit-ema-editor/edit-ema-editor.component.html`](diffhunk://#diff-7f93b3d5700edd531f6ee9f65ed695175ee686d7b1cfa63b46de3669e65fb1efR42):
Added the `sandbox` attribute to the iframe to enhance security by
restricting the actions the iframe can perform.
*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/store/features/editor/withEditor.ts`](diffhunk://#diff-86e692578757ed7f4f6cba5d0aeb07641312f3b17885825d1a45987153ae87f0L53-R55):
Modified the iframe URL generation to avoid caching issues by returning
a new string reference, ensuring the iframe reloads correctly.



## Screenshot



https://github.com/user-attachments/assets/31f15b34-a547-48bf-a6d5-13721fc8541e