chore(UVE): Enhance UX on Personalization workflow (#31125)
This pull request includes several changes to improve the functionality
and user experience of the `dot-ema-editor` components, as well as
updates to the tests to cover new scenarios. The most important changes
include updates to the `DotUveToolbarComponent` and
`EmaContentletToolsComponent` components, as well as modifications to
the SCSS files and test cases.

### Component Updates:

*
[`dot-uve-toolbar.component.ts`](diffhunk://#diff-217a9e619d6590c4f652e85353b9637ba5e464ddeb0424be35aef39bb8dceb30L234-R254):
Added error handling for the personalization confirmation process. If
the process fails, an error message is displayed, and the persona
selector is reset.

*
[`ema-contentlet-tools.component.ts`](diffhunk://#diff-08f74ff61175a0977e14d11a274497b21130e0ea56fdec99b636c0e8207f477dR52):
Added a new input property `disableDeleteButton` to control the state of
the delete button. The button is disabled and shows a tooltip when
`disableDeleteButton` is set.

### SCSS Updates:

*
[`dotcms-theme/components/_dialog.scss`](diffhunk://#diff-123bc2019d92d71d5d2d93da360b73b30eefd6efb45c2c4686e92f6ed2995c68L61-R61):
Changed the `word-break` property from `break-all` to `break-word` for
better text handling.

*
[`dot-ema-shell.component.scss`](diffhunk://#diff-47e09a72b73f6fe223ca726339f5e7a2370d482440adedc9c97af2bea9272b08L44-L47):
Removed unnecessary padding styles from the `.p-dialog-footer` class.

### Test Updates:

*
[`dot-uve-toolbar.component.spec.ts`](diffhunk://#diff-3eaa147616a5d1ff374a5fa27b0f38f0159a9039ef7e8d672dec43631f48a9e1L191-R224):
Added new test cases to cover the error handling in the personalization
process and the use of the `personaEventMock` for consistent test data.
[[1]](diffhunk://#diff-3eaa147616a5d1ff374a5fa27b0f38f0159a9039ef7e8d672dec43631f48a9e1L191-R224)
[[2]](diffhunk://#diff-3eaa147616a5d1ff374a5fa27b0f38f0159a9039ef7e8d672dec43631f48a9e1L483-R544)

*
[`ema-contentlet-tools.component.spec.ts`](diffhunk://#diff-b2ac9d0e3ff7999e99ce6e62e37794a97a1028aa67f6c57ede85e216ab80b388R291-R304):
Added test cases to verify the behavior of the delete button based on
the `disableDeleteButton` input property.

### Store and Mock Updates:

*
[`models.ts`](diffhunk://#diff-a5d9d93bd678484a7d776d09685fcb638de6d969ac141586a1104ba5e744556bR74):
Added `disableDeleteButton` to the `EditorProps` interface to support
the new functionality in the `EmaContentletToolsComponent`.

*
[`withEditor.spec.ts`](diffhunk://#diff-48915538475425e5656151bb2a73df7a99b4c80c2817c085f96352ef3252d5cbL845-R873):
Updated tests to include scenarios where the `disableDeleteButton`
property is set based on the number of contents and the persona state.
[[1]](diffhunk://#diff-48915538475425e5656151bb2a73df7a99b4c80c2817c085f96352ef3252d5cbL845-R873)
[[2]](diffhunk://#diff-48915538475425e5656151bb2a73df7a99b4c80c2817c085f96352ef3252d5cbL860-R889)

## Screenshot


https://github.com/user-attachments/assets/c49cc2ed-b01f-46c4-9658-ae58c9ead283