Add a preview button to Edit Contentlet (#31112)
### Proposed Changes
* Add a preview button to Edit Contentlet

This pull request introduces a new feature to show a preview link in the
content editing form, along with the necessary styling and tests. The
most important changes include adding the preview button, updating the
styles, and writing tests for the new functionality.

### New Feature: Preview Link

* Added a preview button in the `dot-edit-content-form.component.html`
that appears conditionally based on the `$showPreviewLink` property.
* Defined the `$showPreviewLink` computed property to determine whether
the preview link should be shown based on the contentlet's properties.
* Implemented the `showPreview` method to open the preview URL in a new
browser tab.

### Styling

* Added CSS styles for the preview link button in
`dot-edit-content-form.component.scss` to ensure proper alignment and
appearance.

### Testing

* Added tests in `dot-edit-content-form.component.spec.ts` to verify the
rendering and functionality of the preview button, including cases with
and without a URL map.

### Utility Function

* Created a `generatePreviewUrl` function in `functions.util.ts` to
generate the preview URL based on the contentlet's properties.

### Miscellaneous

* Updated imports in `dot-edit-content-form.component.ts` to include the
new utility function and `DotMessagePipe`.
[[1]](diffhunk://#diff-52fdc67adc70be54cf6c07ee18c9340c26cbf50a4d8fe5ed14a21d33def30a23L27-R27)
[[2]](diffhunk://#diff-52fdc67adc70be54cf6c07ee18c9340c26cbf50a4d8fe5ed14a21d33def30a23L41-R45)
[[3]](diffhunk://#diff-52fdc67adc70be54cf6c07ee18c9340c26cbf50a4d8fe5ed14a21d33def30a23L78-R83)

### Checklist
- [x] Tests
- [ ] Translations
- [ ] Security Implications Contemplated (add notes if applicable)